refactor: make styled paper reusable
