+++ packages/client/src/components/StyledPaper.tsx
@@ -1,0 +1,16 @@
+import { Paper, PaperProps, styled } from '@mui/material';
+import { FC, ReactNode } from 'react';
+
+const NewPaper = styled(Paper)(({ theme }) => ({
+  backgroundColor: theme.palette.mode === 'light' ? 'white' : theme.palette.grey['800'],
+  width: '100%',
+  mb: 2,
+  padding: '1rem'
+}));
+
+interface StyledPaperProps extends PaperProps {
+  children: ReactNode[];
+}
+export const StyledPaper: FC<StyledPaperProps> = ({ children, ...otherOptions }) => {
+  return <NewPaper {...otherOptions}>{children}</NewPaper>;
+};
